Falling Platform Finish
Fully implemented the falling platform. Once the player lands on the platform, it will fall for x amount of seconds and then move back up once the player has left the platform. In addition to that, the player will teleport back to the beginning if they stay on the pad while it enters the water.